Demo2

+++ app/src/main/java/com/example/appweather/Model/Main.java
@@ -2,7 +2,7 @@ package com.example.appweather.Model;
 
 public class Main {
     private double temp ;
-    private double pressure ;
+    private float pressure ;
     private int humidity;
     private double temp_min;
     private double temp_max ;
@@ -12,7 +12,6 @@ public class Main {
     public Main() {
     }
 
-
     public double getTemp() {
         return temp;
     }
@@ -21,11 +20,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public double getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(double pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
